+++ ipx800/__init__.py
@@ -1,5 +1,3 @@
-# -*- coding: utf-8 *-*
-
 """Module to control the IPX800 V4 device from GCE Electronics."""
 
 from ipx800.ipx800 import ApiError, IPX800 as ipx800  # noqa